Added Add Drink and Drink Java funtions
